requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anthropic==0.68.0
anyio==4.11.0
asyncer==0.0.8
attrs==25.3.0
backoff==2.2.1
bidict==0.23.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
chainlit==2.8.0
charset-normalizer==3.4.3
chevron==0.14.0
click==8.3.0
colorama==0.4.6
cuid==0.4
dataclasses-json==0.6.7
ddgs==9.6.0
Deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
fastapi==0.116.2
filelock==3.19.1
filetype==1.2.0
frozenlist==1.7.0
fsspec==2025.9.0
google-ai-generativelanguage==0.7.0
google-api-core==2.25.1
google-auth==2.40.3
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.75.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.1
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
inflection==0.5.1
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lance-namespace==0.0.6
lance-namespace-urllib3-client==0.0.14
lancedb==0.25.0
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.76
langchain-experimental==0.3.4
langchain-google-genai==2.1.12
langchain-huggingface==0.3.1
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.30
Lazify==0.4.0
literalai==0.1.201
lxml==6.0.2
MarkupSafe==3.0.2
marshmallow==3.26.1
mcp==1.14.1
monotonic==1.6
mpmath==1.3.0
multidict==6.6.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.5
numpy==2.3.3
openai==1.108.1
opentelemetry-api==1.37.0
opentelemetry-exporter-otlp==1.37.0
opentelemetry-exporter-otlp-proto-common==1.37.0
opentelemetry-exporter-otlp-proto-grpc==1.37.0
opentelemetry-exporter-otlp-proto-http==1.37.0
opentelemetry-instrumentation==0.58b0
opentelemetry-instrumentation-alephalpha==0.47.3
opentelemetry-instrumentation-anthropic==0.47.3
opentelemetry-instrumentation-bedrock==0.47.3
opentelemetry-instrumentation-chromadb==0.47.3
opentelemetry-instrumentation-cohere==0.47.3
opentelemetry-instrumentation-crewai==0.47.3
opentelemetry-instrumentation-google-generativeai==0.47.3
opentelemetry-instrumentation-groq==0.47.3
opentelemetry-instrumentation-haystack==0.47.3
opentelemetry-instrumentation-lancedb==0.47.3
opentelemetry-instrumentation-langchain==0.47.3
opentelemetry-instrumentation-llamaindex==0.47.3
opentelemetry-instrumentation-logging==0.58b0
opentelemetry-instrumentation-marqo==0.47.3
opentelemetry-instrumentation-mcp==0.47.3
opentelemetry-instrumentation-milvus==0.47.3
opentelemetry-instrumentation-mistralai==0.47.3
opentelemetry-instrumentation-ollama==0.47.3
opentelemetry-instrumentation-openai==0.47.3
opentelemetry-instrumentation-openai-agents==0.47.3
opentelemetry-instrumentation-pinecone==0.47.3
opentelemetry-instrumentation-qdrant==0.47.3
opentelemetry-instrumentation-redis==0.58b0
opentelemetry-instrumentation-replicate==0.47.3
opentelemetry-instrumentation-requests==0.58b0
opentelemetry-instrumentation-sagemaker==0.47.3
opentelemetry-instrumentation-sqlalchemy==0.58b0
opentelemetry-instrumentation-threading==0.58b0
opentelemetry-instrumentation-together==0.47.3
opentelemetry-instrumentation-transformers==0.47.3
opentelemetry-instrumentation-urllib3==0.58b0
opentelemetry-instrumentation-vertexai==0.47.3
opentelemetry-instrumentation-watsonx==0.47.3
opentelemetry-instrumentation-weaviate==0.47.3
opentelemetry-instrumentation-writer==0.47.3
opentelemetry-proto==1.37.0
opentelemetry-sdk==1.37.0
opentelemetry-semantic-conventions==0.58b0
opentelemetry-semantic-conventions-ai==0.4.13
opentelemetry-util-http==0.58b0
orjson==3.11.3
ormsgpack==1.10.0
overrides==7.7.0
packaging==25.0
pillow==11.3.0
posthog==3.25.0
primp==0.15.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
PyJWT==2.10.1
pylance==0.36.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-multipart==0.0.20
python-socketio==5.13.0
pywin32==311
PyYAML==6.0.2
referencing==0.36.2
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
SQLAlchemy==2.0.43
sse-starlette==3.0.2
starlette==0.48.0
sympy==1.14.0
syncer==2.0.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
tomli==2.2.1
tools==1.0.5
torch==2.8.0
tqdm==4.67.1
traceloop-sdk==0.47.3
transformers==4.56.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
watchfiles==0.24.0
wrapt==1.17.3
wsproto==1.2.0
xxhash==3.5.0
yarl==1.20.1
zipp==3.23.0
zstandard==0.25.0
